Updated data prep, cleaning, and storage.
Instead of writing individual files to a folder for each general election candidate, those data are now written to a local SQL database. The updated .py file reflects this, and 105 files for Texas were removed as they are no longer necessary.

Additional improvements will be made in the short-term, including moving the output of web scrapes to a SQL table. The string processing at the top of the .py file will be moved to the web scaping script, and a new section loading those data into the tracking_twitter_followers.py file will be written.